transaction_data: update code architecture for clean code, SOLID and factory method implementation
